Hack Computer
Finished construction of the Hack Computer for the HDL project. This entailed integrating the previously designed ALU, memory modules (including RAM16, RAM8, Register, and Bit), and a new CPU component into a comprehensive computer system. Thorough testing has been conducted to ensure the CPU correctly interfaces with the ALU and Memory, and the Hack Computer